rename python init container image name & move AuthConfig to types
