+++ pkg/containerimagebuilderpusher/registryhelpers/cloudregistry.go
@@ -39,16 +39,6 @@ const (
 	providerCredentialsFileName = "credentials"
 )
 
-// AuthConfig carries the images and provider secret the auth/merge init containers need.
-type AuthConfig struct {
-	AWSCLIImage                string
-	AzureCLIImage              string
-	GCloudCLIImage             string
-	RegistryProviderSecretName string
-	PythonImage                string
-	PythonImagePullPolicy      string
-}
-
 // CloudRegistryHelper authenticates to one cloud registry vendor.
 type CloudRegistryHelper interface {
 	// Matches reports whether host belongs to this vendor.
